improv: check to ensure entity is touching caelgae bounding box

+++ src/main/java/net/zepalesque/redux/block/natural/CaelgaeBlock.java
@@ -91,11 +91,16 @@ public class CaelgaeBlock extends Block implements BonemealableBlock {
 	
 	@Override
 	protected void entityInside(BlockState state, Level level, BlockPos pos, Entity entity) {
-		if (entity.getType().equals(EntityType.FALLING_BLOCK)) {
-			level.destroyBlock(pos, true);
-		} else if (entity instanceof LivingEntity && !entity.getType().is(EntityTypeTags.AQUATIC)) {
+		if (entity.getType().equals(EntityType.FALLING_BLOCK)) level.destroyBlock(pos, true);
+		else if (entity instanceof LivingEntity && !entity.getType().is(EntityTypeTags.AQUATIC) && isEntityIntersecting(state, level, pos, entity))
 			entity.setDeltaMovement(entity.getDeltaMovement().multiply(new Vec3(0.85, 0.85, 0.85)));
-		}
+	}
+	
+	protected boolean isEntityIntersecting(BlockState state, Level level, BlockPos pos, Entity entity) {
+		var entityBox = entity.getBoundingBox();
+		var selfBox = getShape(state, level, pos, CollisionContext.of(entity)).bounds().move(pos);
+		
+		return entityBox.intersects(selfBox);
 	}
 	
 	@Override
